*Fixed Herobrine statue not dropping an item when the bottom half is mined in survival mode *Minor optimization *Fixed Herobrine loot table not specifying a namespace for type key which can cause issues with many mods *Herobrine Statue loot table now specifies rolls as an integer and the loot pool is named *Bumped version to 0.6.5-preview-build-2

+++ src/main/java/com/herobrinemod/herobrine/entities/HerobrineSpyEntity.java
@@ -33,7 +33,7 @@ public class HerobrineSpyEntity extends HerobrineEntity {
         this.goalSelector.add(0, new SwimGoal(this));
         this.goalSelector.add(1, new LookAtEntityGoal(this, PlayerEntity.class, 1024.0f));
         //this.goalSelector.add(2, new LookAtEntityGoal(this, SurvivorEntity.class, 1024.0f));
-        this.goalSelector.add(3, new FleeEntityGoal<>(this, PlayerEntity.class, (float) ConfigHandler.herobrineConfig.readDouble("HerobrineSpyObservationDistance"), 0.7, 1.0));
+        this.goalSelector.add(3, new FleeEntityGoal<>(this, PlayerEntity.class, ConfigHandler.herobrineConfig.readInt("HerobrineSpyObservationDistance"), 0.7, 1.0));
         //this.goalSelector.add(4, new FleeEntityGoal<>(this, SurvivorEntity.class, 32.0f, 0.7, 1.0));
     }
